started the outline
started a glossary of terms. please feel free to adjust or add to the list of terms. and if you want to pick another category idea and add to the outline, please do! whatever ends up in this outline I will add to the photoshop document that will be used to create our final product. right now i'm just throwing this at the wall and seeing what sticks, so please adjust or let me know what ideas you have!